Rename codegen files
