+++ apps/parsley/package.json
@@ -15,7 +15,7 @@
     "build:profiler:staging": "PROFILER=true PROFILE_HEAD=\"<script src=\"https://localhost:8097\"></script>\" yarn build:staging",
     "check-types": "tsc -p tsconfig.json --noEmit",
     "clean": "rm -rf node_modules && rm -rf build && yarn install",
-    "codegen": "graphql-codegen --config codegen.ts",
+    "codegen": "graphql-codegen --config graphql.config.ts",
     "cy:open": "yarn evg-db-ops --reseed-and-dump; cypress open",
     "cy:run": "yarn evg-db-ops --reseed-and-dump; cypress run",
     "deploy:beta": "env-cmd -e beta yarn deploy-utils-deploy",
