+++ apps/spruce/package.json
@@ -16,7 +16,7 @@
     "check-schema-and-codegen": "env-cmd -e local vite-node scripts/check-schema-and-codegen/script.ts",
     "check-types": "tsc -p tsconfig.json --noEmit",
     "clean": "rm -rf node_modules && rm -rf dist && yarn install",
-    "codegen": "graphql-codegen-esm --config codegen.ts",
+    "codegen": "graphql-codegen-esm --config graphql.config.ts",
     "cy:open": "yarn evg-db-ops --reseed-and-dump; cypress open",
     "cy:run": "yarn evg-db-ops --reseed-and-dump; cypress run",
     "cy:test": "yarn evg-db-ops --reseed-and-dump; cypress run --spec",
